+++ src/System.Threading/tests/CountdownEventTests.cs
@@ -1,15 +1,10 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Xunit;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using System.Diagnostics;
+using Xunit;
 
-namespace Test
+namespace System.Threading.Tests
 {
     public class CountdownEventTests
     {
