+++ src/System.Threading/tests/ReaderWriterLockSlimTests.cs
@@ -1,12 +1,10 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace Test
+namespace System.Threading.Tests
 {
     public static class ReaderWriterLockSlimTests
     {
